Add basic signals for enemy events Change power-up drop change field to range
