+++ scenes/enemies/base_enemy.tscn
@@ -11,3 +11,5 @@ script = ExtResource("1_nxebk")
 
 [node name="CollisionShape2D" type="CollisionShape2D" parent="."]
 shape = SubResource("CapsuleShape2D_wmcpu")
+
+[node name="Sprite2D" type="Sprite2D" parent="."]
